Minor cleanups
