+++ pynicotine/gtkgui/downloads.py
@@ -24,6 +24,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
+import re
 
 from gi.repository import Gdk
 from gi.repository import Gtk
@@ -38,6 +39,7 @@ from pynicotine.gtkgui.utils import human_speed
 from pynicotine.gtkgui.utils import open_file_path
 from pynicotine.gtkgui.utils import PopupMenu
 from pynicotine.gtkgui.utils import set_treeview_selected_row
+from pynicotine.logfacility import log
 
 
 class Downloads(TransferList):
@@ -90,6 +92,8 @@ class Downloads(TransferList):
         frame.ExpandDownloads.connect("toggled", self.on_expand_downloads)
         frame.ExpandDownloads.set_active(frame.np.config.sections["transfers"]["downloadsexpanded"])
 
+        self.update_download_filters()
+
     def on_try_clear_queued(self, widget):
         option_dialog(
             parent=self.frame.MainWindow,
@@ -122,6 +126,56 @@ class Downloads(TransferList):
             # Group by folder, show user folders in collapsed mode
 
             self.frame.DownloadList.expand_to_path(user_path)
+
+    def update_download_filters(self):
+
+        proccessedfilters = []
+        outfilter = "(\\\\("
+        failed = {}
+        df = sorted(self.frame.np.config.sections["transfers"]["downloadfilters"])
+        # Get Filters from config file and check their escaped status
+        # Test if they are valid regular expressions and save error messages
+
+        for item in df:
+            dfilter, escaped = item
+            if escaped:
+                dfilter = re.escape(dfilter)
+                dfilter = dfilter.replace("\\*", ".*")
+
+            try:
+                re.compile("(" + dfilter + ")")
+                outfilter += dfilter
+                proccessedfilters.append(dfilter)
+            except Exception as e:
+                failed[dfilter] = e
+
+            proccessedfilters.append(dfilter)
+
+            if item is not df[-1]:
+                outfilter += "|"
+
+        # Crop trailing pipes
+        while outfilter[-1] == "|":
+            outfilter = outfilter[:-1]
+
+        outfilter += ")$)"
+        try:
+            re.compile(outfilter)
+            self.frame.np.config.sections["transfers"]["downloadregexp"] = outfilter
+            # Send error messages for each failed filter to log window
+            if len(failed) >= 1:
+                errors = ""
+
+                for dfilter, error in failed.items():
+                    errors += "Filter: %s Error: %s " % (dfilter, error)
+
+                error = _("Error: %(num)d Download filters failed! %(error)s ", {'num': len(failed), 'error': errors})
+                log.add(error)
+
+        except Exception as e:
+            # Strange that individual filters _and_ the composite filter both fail
+            log.add(_("Error: Download Filter failed! Verify your filters. Reason: %s", e))
+            self.frame.np.config.sections["transfers"]["downloadregexp"] = ""
 
     def on_expand_downloads(self, widget):
