+++ pynicotine/gtkgui/frame.py
@@ -23,7 +23,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
-import re
 import signal
 import sys
 import threading
@@ -68,7 +67,6 @@ from pynicotine.gtkgui.utils import scroll_bottom
 from pynicotine.gtkgui.utils import TextSearchBar
 from pynicotine.gtkgui.utils import update_widget_visuals
 from pynicotine.logfacility import log
-from pynicotine.nowplaying import NowPlaying
 from pynicotine.pynicotine import NetworkEventProcessor
 from pynicotine.utils import get_latest_version
 from pynicotine.utils import make_version
@@ -340,10 +338,6 @@ class NicotineFrame:
 
         self.statistics = Statistics(self, self.np.config)
 
-        """ Now Playing """
-
-        self.now_playing = NowPlaying(self.np.config)
-
         """ Connect """
 
         if self.np.config.need_config():
@@ -357,13 +351,6 @@ class NicotineFrame:
             self.on_connect()
 
         self.update_bandwidth()
-
-        """ Search """
-
-        self.UserSearchCombo.hide()
-        self.RoomSearchCombo.hide()
-
-        self.update_download_filters()
 
         """ Tab Signals """
 
@@ -1893,55 +1880,6 @@ class NicotineFrame:
 
         self.RoomSearchCombo.set_sensitive(act)
 
-    def update_download_filters(self):
-        proccessedfilters = []
-        outfilter = "(\\\\("
-        failed = {}
-        df = sorted(self.np.config.sections["transfers"]["downloadfilters"])
-        # Get Filters from config file and check their escaped status
-        # Test if they are valid regular expressions and save error messages
-
-        for item in df:
-            dfilter, escaped = item
-            if escaped:
-                dfilter = re.escape(dfilter)
-                dfilter = dfilter.replace("\\*", ".*")
-
-            try:
-                re.compile("(" + dfilter + ")")
-                outfilter += dfilter
-                proccessedfilters.append(dfilter)
-            except Exception as e:
-                failed[dfilter] = e
-
-            proccessedfilters.append(dfilter)
-
-            if item is not df[-1]:
-                outfilter += "|"
-
-        # Crop trailing pipes
-        while outfilter[-1] == "|":
-            outfilter = outfilter[:-1]
-
-        outfilter += ")$)"
-        try:
-            re.compile(outfilter)
-            self.np.config.sections["transfers"]["downloadregexp"] = outfilter
-            # Send error messages for each failed filter to log window
-            if len(failed) >= 1:
-                errors = ""
-
-                for dfilter, error in failed.items():
-                    errors += "Filter: %s Error: %s " % (dfilter, error)
-
-                error = _("Error: %(num)d Download filters failed! %(error)s ", {'num': len(failed), 'error': errors})
-                log.add(error)
-
-        except Exception as e:
-            # Strange that individual filters _and_ the composite filter both fail
-            log.add(_("Error: Download Filter failed! Verify your filters. Reason: %s", e))
-            self.np.config.sections["transfers"]["downloadregexp"] = ""
-
     def on_search(self, widget):
         self.searches.on_search()
 
@@ -2624,7 +2562,7 @@ class NicotineFrame:
             self.searches.maxstoredresults = config["searches"]["max_stored_results"]
 
         # Modify GUI
-        self.update_download_filters()
+        self.downloads.update_download_filters()
         self.np.config.write_configuration()
 
         if not config["ui"]["trayicon"] and self.tray.is_tray_icon_visible():
