+++ pynicotine/pynicotine.py
@@ -41,6 +41,7 @@ from pynicotine import transfers
 from pynicotine.config import Config
 from pynicotine.geoip.ip2location import IP2Location
 from pynicotine.logfacility import log
+from pynicotine.nowplaying import NowPlaying
 from pynicotine.pluginsystem import PluginHandler
 from pynicotine.shares import Shares
 from pynicotine.slskmessages import new_id
@@ -134,6 +135,7 @@ class NetworkEventProcessor:
         self.statistics = Statistics(self.config, self.ui_callback)
         self.shares = Shares(self, self.config, self.queue, self.ui_callback)
         self.pluginhandler = PluginHandler(self.ui_callback, plugins, self.config)
+        self.now_playing = NowPlaying(self.config)
 
         script_dir = os.path.dirname(__file__)
         file_path = os.path.join(script_dir, "geoip/ipcountrydb.bin")
